fix: Fix mypy type checking errors for CI pipeline
- Add proper type annotations throughout the codebase
- Fix implicit Optional types with explicit Optional annotations
- Add None checks for all Optional type usages
- Fix SQLAlchemy model type annotations
- Fix Collection type usage and dictionary assignments
- Ensure all return types are properly annotated

This resolves all mypy errors blocking the CI pipeline.

+++ src/rag_processor/chunker.py
@@ -106,7 +106,7 @@ class TextChunkingStrategy(ChunkingStrategy):
             return len(self.encoding.encode(text))
         else:
             # Fallback to word count approximation
-            return len(text.split()) * 1.3
+            return int(len(text.split()) * 1.3)
     
     def _create_chunk(
         self,
